==> analysis/01_analysis-stock-unweighted.qmd <==
---
title: "AEA Formal Privacy Questionnaire Analysis"
author-title: "Authors"
authors: "Aaron R. Williams"
affiliation: "Urban Institute"
date: today
format:
  html:
    theme: www/urbn.scss
    self-contained: true
    code-fold: true
    code-line-numbers: true
    html-math-method: katex
    df-print: paged
    toc: true
    toc-depth: 3
    number-sections: true
    number-depth: 3
    highlight-style: pygments
#bibliography: references.bib
execute: 
  warning: false
  message: false
  fig-width: 6.5
  fig-hieght: 4
editor_options: 
  chunk_output_type: console
---

```{r opts, echo = FALSE}
knitr::opts_chunk$set(
  fig.path = "../figures/"
)
```

```{r setup}
options(scipen = 999)

library(tidyverse)
library(qualtRics)
library(urbnthemes)

set_urbn_defaults(style = "print")

```

```{r results = "hide"}
surveys <- all_surveys() 

aea <- fetch_survey(surveyID = surveys$id[1])

aea <- aea %>%
  filter(DistributionChannel != "preview")

```

## Response Quality

Most IP addresses are unique. 

```{r}
aea %>%
  count(IPAddress) %>%
  count(n_frequency = n) %>%
  mutate(prop = (n_frequency * n) / sum(n_frequency * n))

```

```{r}
aea %>%
  count(LocationLongitude, LocationLatitude) %>%
  count(n_frequency = n) %>%
  mutate(prop = (n_frequency * n) / sum(n_frequency * n))

aea %>%
  filter(!is.na(LocationLongitude) & !is.na(LocationLatitude)) %>%
  count(LocationLongitude, LocationLatitude) %>%
  count(n_frequency = n) %>%
  mutate(prop = (n_frequency * n) / sum(n_frequency * n))

```

419 responses have missing longitude and latitude. 

### Responses by Day

```{r response-plot}
aea %>%
  mutate(id = rank(StartDate)) %>%
  ggplot() + 
  geom_linerange(aes(xmin = StartDate, xmax = EndDate, y = id)) +
  geom_vline(
    aes(xintercept = lubridate::ymd_hms("2022-04-25 13:00:00")), color = "red"
  ) +
  geom_vline(
    aes(xintercept = lubridate::ymd_hms("2022-05-09 13:00:00")),
    color = "red"
  ) + 
  annotate(
    geom = "text", 
    x = lubridate::ymd_hms("2022-04-26 18:00:00"),
    y = 1100, 
    label = "Email 1"
  ) +
  annotate(
    geom = "text", 
    x = lubridate::ymd_hms("2022-05-10 18:00:00"),
    y = 1100, 
    label = "Email 2"
  ) +  
  labs(
    title = "The Median Response Time Was About Six Minutes and Responses Were\nConcentrated After Emails",
    y = NULL
  ) +
  remove_axis("y") +
  theme(axis.text.y = element_blank())

```

### Distribution of Durations

```{r}
aea %>%
  mutate(duration = `Duration (in seconds)` / 60) %>%
  summarize(
    mean(duration),
    median(duration)
  )

```

### Response Progress

```{r}
count(aea, Progress)

```

* Most responses were more than 90% complete

## Section 1

### Question 1

1.1: Do you reside in the United States?

```{r q1.1_unweighted}
aea %>%
  count(Q1_1) %>%
  mutate(Q1_1 = as.character(Q1_1)) %>%
  filter(!is.na(Q1_1)) %>%
  mutate(Q1_1 = factor(Q1_1, levels = c("Yes", "No"))) %>%
  ggplot(aes(Q1_1, n)) +
  geom_col() +
  scale_y_continuous(expand = expansion(mult = c(0, 0.2))) +
  labs(
    title = "Do you reside in the United States?",
    x = NULL,
    y = NULL,
    caption = "10 respondents did not answer"
  ) +
  remove_ticks()

```

```{r}
aea %>%
  count(Q1_1) %>%
  mutate(Q1_1 = as.character(Q1_1)) %>%
  filter(!is.na(Q1_1)) %>%
  mutate(prop = n / sum(n))

```

1.2: Are you currently a student?

```{r q2.2_unweighted}
aea %>%
  count(Q1_2) %>%
  mutate(Q1_2 = as.character(Q1_2)) %>%
  filter(!is.na(Q1_2)) %>%
  mutate(Q1_2 = factor(Q1_2, levels = c("Yes", "No"))) %>%
  ggplot(aes(Q1_2, n)) +
  geom_col() +
  scale_y_continuous(expand = expansion(mult = c(0, 0.2))) +
  labs(
    title = "Are You Currently a Student?",
    x = NULL,
    y = NULL,
    caption = "11 respondents did not answer"
  ) +
  remove_ticks()

```


```{r}
aea %>%
  count(Q1_2) %>%
  mutate(Q1_2 = as.character(Q1_2)) %>%
  filter(!is.na(Q1_2)) %>%
  mutate(prop = n / sum(n))

```

1.3: What is the highest degree or level of school you have completed?

```{r q1.3_unweighted}
aea %>%
  count(Q1_3) %>%
  slice(1:5) %>%
  mutate(Q1_3 = fct_recode(
    Q1_3, 
    `High school or General\nEducational Development (GED)` = 
      "High school or General Educational Development (GED)")
  ) %>%  
  ggplot(aes(n, Q1_3)) +
  geom_col() +
  scale_x_continuous(expand = expansion(mult = c(0, 0.2))) +
  labs(
    title = "What is the highest degree or level of school you have completed?",
    x = NULL,
    y = NULL,
    caption = "11 respondents did not answer"
  ) +
  theme(panel.grid = element_blank())

```

```{r}
aea %>%
  count(Q1_3) %>%
  mutate(Q1_3 = as.character(Q1_3)) %>%
  filter(!is.na(Q1_3)) %>%
  mutate(prop = n / sum(n))

```

1.4: In what year did you finish your highest degree or level of school?

```{r q1.4_unweighted}
aea %>%
  ggplot(aes(Q1_4)) +
  geom_bar(width = 5) +
  scale_y_continuous(expand = expansion(mult = c(0, 0.2))) +
  labs(
    title = "In what year did you finish your highest degree or level of school?",
    x = NULL,
    y = NULL,
    caption = "20 respondents did not answer"
  )

```

1.5: Which of the following best describes your most recent primary employer?

```{r q1.5_unweighted}
aea %>%
  count(Q1_5) %>%
  slice(1:5) %>%
  ggplot(aes(n, Q1_5)) +
  geom_col() +
  scale_x_continuous(expand = expansion(mult = c(0, 0.2))) +
  labs(
    title = "Which of the following best describes your most recent primary employer?",
    x = NULL,
    y = NULL,
    caption = "12 respondents did not answer"
  ) +
  theme(panel.grid = element_blank())

```

```{r}
aea %>%
  count(Q1_5) %>%
  mutate(Q1_5 = as.character(Q1_5)) %>%
  filter(!is.na(Q1_5)) %>%
  mutate(prop = n / sum(n))

```

1.6: Have you refereed at least one peer-reviewed journal article in the past five years?

```{r q1.6_unweighted}
aea %>%
  count(Q1_6) %>%
  slice(1:2) %>%
  ggplot(aes(Q1_6, n)) +
  geom_col() +
  scale_y_continuous(expand = expansion(mult = c(0, 0.2))) +
  labs(
    title = "Have you refereed at least one peer-reviewed journal article in the past five years?",
    x = NULL,
    y = NULL,
    caption = "12 respondents did not answer"
  ) +
  remove_ticks()

```

```{r}
aea %>%
  count(Q1_6) %>%
  mutate(Q1_6 = as.character(Q1_6)) %>%
  filter(!is.na(Q1_6)) %>%
  mutate(prop = n / sum(n))

```

1.7: Select up to three Journal of Economic Literature (JEL) Classification Codes that best describe your area of work:

* Most popular codes
* Patterns in codes

```{r}
jel <- aea %>%
  select(starts_with("Q1_7"))

top_jel <- jel %>%
  pivot_longer(everything(), names_to = "a", values_to = "JEL code") %>%
  count(`JEL code`) %>%
  filter(!is.na(`JEL code`))

```


```{r q1.7_top-ten}
top_jel %>%
  slice_max(n, n = 10) %>%
  mutate(`JEL code` = fct_recode(
    `JEL code`,
    `O. Economic Development, Innovation,\nTechnological Change, and Growth` =
      "O. Economic Development, Innovation, Technological Change, and Growth")
  ) %>%
  mutate(`JEL code` = fct_reorder(`JEL code`, n)) %>%
  ggplot(aes(n, `JEL code`)) +
  geom_col() +
  scale_x_continuous(expand = expansion(mult = c(0, 0.2))) +
  labs(
    title = "Select up to Three Journal of Economic Literature (JEL) Classification Codes that\nBest Describe Your Area of Work",
    x = NULL,
    y = NULL
  ) +
  theme(panel.grid = element_blank())

```

```{r q1.7, fig.height = 5}
top_jel %>%
  mutate(`JEL code` = fct_recode(
    `JEL code`,
    `M. Business Administration and Business Economics;\nMarketing; Accounting; Personnel Economics` = 
      "M. Business Administration and Business Economics; Marketing; Accounting; Personnel Economics",
    `O. Economic Development, Innovation,\nTechnological Change, and Growth` =
      "O. Economic Development, Innovation, Technological Change, and Growth",
    `Q. Agricultural and Natural Resource Economics;\nEnvironmental and Ecological Economics` =
      "Q. Agricultural and Natural Resource Economics; Environmental and Ecological Economics")
  ) %>%
  mutate(`JEL code` = fct_reorder(`JEL code`, n)) %>%
  ggplot(aes(n, `JEL code`)) +
  geom_col() +
  scale_x_continuous(expand = expansion(mult = c(0, 0.2))) +
  labs(
    title = "Select up to Three Journal of Economic Literature (JEL) Classification Codes that\nBest Describe Your Area of Work",
    x = NULL,
    y = NULL
  ) +
  theme(panel.grid = element_blank())

```

## Section 2

### Question 1 {.tabset}

Q2_1 - 2.1: Please rate the following methods based on how frequently you have used the method on cross-sectional data in the past year:

#### Visualization

```{r q2.1_unweighed}
responses <- c("Always", "Frequently")

aea %>%
  summarize(
    `Multiple linear regression` = mean(Q2_1_1 %in% responses, na.rm = TRUE),
    `Weighted multiple linear regression` = mean(Q2_1_2 %in% responses, na.rm = TRUE),
    `Tobit` = mean(Q2_1_3 %in% responses, na.rm = TRUE),
    `Probit/logit regression` = mean(Q2_1_4 %in% responses, na.rm = TRUE),
    `Regression discontinuity or kink design` = mean(Q2_1_5 %in% responses, na.rm = TRUE),
    `Merging multiple data sources` = mean(Q2_1_6 %in% responses, na.rm = TRUE),
    `Two Stage Least Squares (2SLS)` = mean(Q2_1_7 %in% responses, na.rm = TRUE)
  ) %>%
  pivot_longer(
    everything(),
    names_to = "Method", 
    values_to = "Proportion 'Always' or 'Frequently'"
  ) %>%
  mutate(Method = fct_reorder(Method, `Proportion 'Always' or 'Frequently'`)) %>%
  ggplot(aes(x = `Proportion 'Always' or 'Frequently'`, y = Method)) +
  geom_col() +
  geom_text(aes(label = round(`Proportion 'Always' or 'Frequently'`, 2), hjust = -0.3)) +
  scale_x_continuous(limits = c(0, 0.8)) +
  labs(
    title = "Multiple Linear Regression is the Most Popular Method",
    subtitle = "Proportion of Respondents Rating Each Method with 'Always' or 'Frequently'",
    x = NULL,
    y = NULL,
    caption = "147 respondents did not answer"
  ) +
  theme(panel.grid = element_blank())

```

#### Table

The number observations missing all Q2.1 variables. 

```{r}
count(aea, missing_all = is.na(Q2_1_1) & is.na(Q2_1_2) & is.na(Q2_1_3) & is.na(Q2_1_4) & is.na(Q2_1_5) & is.na(Q2_1_6) & is.na(Q2_1_7))

```

Multiple linear regression

```{r}
count(aea, Q2_1_1)

```

Weighted multiple linear regression

```{r}
count(aea, Q2_1_2)

```

Tobit

```{r}
count(aea, Q2_1_3)

```

Probit/logit regression

```{r}
count(aea, Q2_1_4)

```

Regression discontinuity or kink design

```{r}
count(aea, Q2_1_5)

```

Merging multiple data sources

```{r}
count(aea, Q2_1_6)

```

Two Stage Least Squares (2SLS)

```{r}
count(aea, Q2_1_7)

```

### Question 2

See analysis/02_analysis-text.qmd.

### Question 3 {.tabset}

2.3: Please rate the following information obtained from multiple linear regression based on its importance for your work:

#### Visualization

```{r q2.3_unweighted}
responses <- c("Very important", "Important")

aea %>%
  summarize(
    `Estimated coefficients` = mean(Q2_3_1 %in% responses, na.rm = TRUE),
    `Estimated standard errors, T-statistics for\ncoefficients, or P-values for coefficients` = mean(Q2_3_2 %in% responses, na.rm = TRUE),
    `Adjusted or multiple R-Squared of\nthe regression line` = mean(Q2_3_3 %in% responses, na.rm = TRUE),
    `F-statistic on the regression` = mean(Q2_3_4 %in% responses, na.rm = TRUE),
    `Residuals` = mean(Q2_3_5 %in% responses, na.rm = TRUE),
    `Residual standard error` = mean(Q2_3_6 %in% responses, na.rm = TRUE),
    `Predicted values` = mean(Q2_3_7 %in% responses, na.rm = TRUE),
    `Average predicted values for a given\ncharacteristic or marginal predictions` = mean(Q2_3_8 %in% responses, na.rm = TRUE)
  ) %>%
  pivot_longer(
    everything(),
    names_to = "Method", 
    values_to = "Proportion 'Very important' or 'Important'"
  ) %>%
  mutate(Method = fct_reorder(Method, `Proportion 'Very important' or 'Important'`)) %>%
  ggplot(aes(x = `Proportion 'Very important' or 'Important'`, y = Method)) +
  geom_col() +
  geom_text(aes(label = round(`Proportion 'Very important' or 'Important'`, 2), hjust = -0.3)) +  
  scale_x_continuous(limits = c(0, 0.8)) +
  labs(
    title = "Coefficient-level Information is the Most Important",
    subtitle = "Proportion of Respondents Rating Each Component with 'Very important' or 'Important'",
    x = NULL,
    y = NULL,
    caption = "170 respondents did not answer"
  ) +
  theme(panel.grid = element_blank())

```

#### Table

The number observations missing all Q2.3 variables. 

```{r}
count(aea, missing_all = is.na(Q2_3_1) & is.na(Q2_3_2) & is.na(Q2_3_3) & is.na(Q2_3_4) & is.na(Q2_3_5) & is.na(Q2_3_6) & is.na(Q2_3_7) & is.na(Q2_3_8))

```

Estimated coefficients

```{r}
count(aea, Q2_3_1)

```

Estimated standard errors, T-statistics for coefficients, or P-values for coefficients

```{r}
count(aea, Q2_3_2)

```

Adjusted or multiple R-Squared of the regression line

```{r}
count(aea, Q2_3_3)

```

F-statistic on the regression

```{r}
count(aea, Q2_3_4)

```

Residuals

```{r}
count(aea, Q2_3_5)

```

Residual standard error

```{r}
count(aea, Q2_3_6)

```

Predicted values

```{r}
count(aea, Q2_3_7)

```

Average predicted values for a given characteristic or marginal predictions

```{r}
count(aea, Q2_3_8)

```



### Question 5

See analysis/02_analysis-text.qmd.

## Section 3

### Question 1

Q3_1 - 3.1: Please rate your familiarity with the concept of differential privacy.

```{r q3.1_unweighted}
aea %>%
  count(Q3_1) %>%
  filter(n != 181) %>%
  mutate(Q3_1 = fct_recode(
    Q3_1, 
    `Have heard the term but am\nnot familiar with any details` = 
      "Have heard the term but am not familiar with any details",
    `Have read a blog, newspaper, or non-academic\nreport or discussion on the topic` = 
      "Have read a blog, newspaper, or non-academic report or discussion on the topic",
    `Feel confident implementing\nthese methods on my own` = 
      "Feel confident implementing these methods on my own")
  ) %>%
  mutate(Q3_1 = fct_rev(Q3_1)) %>%
  ggplot(aes(n, Q3_1)) +
  geom_col() +
  geom_text(aes(label = n, hjust = -0.5)) +
  scale_x_continuous(limits = c(NA, 500)) +
  labs(
    title = "Most Respondents Knew Nothing about Differential Privacy",
    subtitle = "Text: Please rate your familiarity with the concept of differential privacy",
    x = NULL,
    y = NULL,
    caption = "181 respondents did not answer"
  ) +
  theme(panel.grid = element_blank())

```

```{r q3.1-primary-employer_unweighted, fig.width=10}
aea %>%
    mutate(
    Q1_5 = fct_recode(
      Q1_5,
      `Industry (For or Not-for profit)` = "For-profit business",
      `Industry (For or Not-for profit)` = "Not-for-profit",
      `Government` = "Federal government",
      `Government` = "State or local government"
    )
  ) %>%
  count(Q3_1, Q1_5) %>%
  filter(
    !is.na(as.character(Q3_1)),
    !is.na(as.character(Q1_5))
  ) %>%
  mutate(Q3_1 = fct_recode(
    Q3_1, 
    `Have heard the term but am\nnot familiar with any details` = 
      "Have heard the term but am not familiar with any details",
    `Have read a blog, newspaper, or non-academic\nreport or discussion on the topic` = 
      "Have read a blog, newspaper, or non-academic report or discussion on the topic",
    `Feel confident implementing\nthese methods on my own` = 
      "Feel confident implementing these methods on my own")
  ) %>%
  mutate(Q3_1 = fct_rev(Q3_1)) %>%
  ggplot(aes(n, Q3_1)) +
  geom_col() +
  geom_text(aes(label = n, hjust = -0.5)) +  
  scale_x_continuous(expand = expansion(mult = c(0, 0.2))) +
  facet_wrap(~Q1_5, scales = "free_x") +
  labs(
    title = "Most Respondents Knew Nothing about Differential Privacy",
    subtitle = "Text: Please rate your familiarity with the concept of differential privacy",
    x = NULL,
    y = NULL,
    caption = "181 respondents did not answer"
  ) +
  theme(
    panel.grid = element_blank(),
    panel.spacing = unit(2, "lines")
  )

```

Calculate proportion with no familiarity with DP.

```{r}
(454 + 206) / (454 + 206 + 105 + 59 + 23)

```

### Question 2

Q3_2 - 3.2: Please rate your familiarity with the concept of formal privacy

```{r q3.2_unweighted}
aea %>%
  count(Q3_2) %>%
  filter(n != 183) %>%
  mutate(Q3_2 = fct_recode(
    Q3_2,
    `Have heard the term in relation to Differential\nPrivacy, but I do not know the difference` =
      "Have heard the term in relation to Differential Privacy, but I do not know the difference",
    `Am familiar with the concept and the distinction\nfrom differential privacy` = 
      "Am familiar with the concept and the distinction from differential privacy")
  ) %>%
  mutate(Q3_2 = fct_rev(Q3_2)) %>%
  ggplot(aes(n, Q3_2)) +
  geom_col() +
  geom_text(aes(label = n, hjust = -0.5)) +  
  scale_x_continuous(limits = c(NA, 600)) +
  labs(
    title = "Most Respondents Knew Nothing about Formal Privacy",
    subtitle = "Text: Please rate your familiarity with the concept of formal privacy",
    x = NULL,
    y = NULL,
    caption = "183 respondents did not answer"
  ) +
  theme(panel.grid = element_blank())

```

Calculate the proportion with no or little experience. 

```{r}
aea %>%
  count(Q3_2)

(560 + 227) / (560 + 227 + 58)

```

### Question 3

Q3_3 - 3.3: What share of economists you know in your professional circles have discussed the US Census Bureau's adoption of differential privacy/formal privacy for the 2020 Census?

```{r q3.3_unweighted}
aea %>%
  count(Q3_3) %>%
  filter(n != 185) %>%
  mutate(Q3_3 = fct_rev(Q3_3)) %>%
  ggplot(aes(n, Q3_3)) +
  geom_col() +
  geom_text(aes(label = n, hjust = -0.5)) +  
  scale_x_continuous(limits = c(NA, 600)) +
  labs(
    title = "Most Economists Have Not Discussed Differential Privacy and the 2020 Census",
    subtitle = "Text: What share of economists you know in your professional circles have discussed the US Census Bureau's\nadoption of differential privacy/formal privacy for the 2020 Census?",
    x = NULL,
    y = NULL,
    caption = "185 respondents did not answer"
  ) +
  theme(panel.grid = element_blank())

```

### Question 4

Q3_4 - 3.4: What share of economists you know in your professional circles work with differential privacy/formal privacy methods?

```{r q3.4_unweighted}
aea %>%
  count(Q3_4) %>%
  filter(n != 193) %>%
  mutate(Q3_4 = fct_rev(Q3_4)) %>%
  ggplot(aes(n, Q3_4)) +
  geom_col() +
  geom_text(aes(label = n, hjust = -0.5)) +  
  scale_x_continuous(limits = c(NA, 610)) +
  labs(
    title = "Most Economists Don't Work with Differential Privacy",
    subtitle = "Text: What share of economists you know in your professional circles work with differential privacy/formal\nprivacy methods?",
    x = NULL,
    y = NULL,
    caption = "193 respondents did not answer"
  ) +
  theme(panel.grid = element_blank())

```

### Question 5

Q3_5 - 3.5: I do not have an opinion on Differential Privacy/formal privacy

```{r q3.5_unweighted}
aea %>%
  count(Q3_5) %>%
  filter(n != 187) %>%
  mutate(Q3_5 = fct_rev(Q3_5)) %>%
  ggplot(aes(n, Q3_5)) +
  geom_col() +
  geom_text(aes(label = n, hjust = -0.5)) +  
  scale_x_continuous(limits = c(NA, 800)) +
  labs(
    title = "Most Respondents Have no Opinion on Differential Privacy",
    subtitle = "Text: I do not have an opinion on Differential Privacy/formal privacy",
    x = NULL,
    y = NULL,
    caption = "187 respondents did not answer"
  ) +
  theme(panel.grid = element_blank())

```

Calculate the proportion who agree. 

```{r}
667 / (667 + 174)


```

### Question 6

Q3_6 - Differential privacy/formal privacy is a new approach to protecting privacy that seems to be mostly popular among computer scientists.

```{r q3.6_unweighted}
aea %>%
  count(Q3_6) %>%
  filter(n != 216) %>%
  mutate(Q3_6 = fct_rev(Q3_6)) %>%
  ggplot(aes(n, Q3_6)) +
  geom_col() +
  geom_text(aes(label = n, hjust = -0.5)) +  
  scale_x_continuous(limits = c(NA, 800)) +
  labs(
    title = "Most Economists think Differential Privacy is Popular with Computer Scientists",
    subtitle = "Text: Differential privacy/formal privacy is a new approach to protecting privacy that seems to be mostly popular\namong computer scientists.",
    x = NULL,
    y = NULL,
    caption = "216 respondents did not answer"
  ) +
  theme(panel.grid = element_blank())

```

### Question 7

Q3_7 - Differential privacy/formal privacy is an approach to protecting privacy that distorts the results more than necessary given the actual risks.

```{r q3.7_unweighted}
aea %>%
  count(Q3_7) %>%
  filter(n != 212) %>%
  mutate(Q3_7 = fct_rev(Q3_7)) %>%
  ggplot(aes(n, Q3_7)) +
  geom_col() +
  geom_text(aes(label = n, hjust = -0.5)) +  
  scale_x_continuous(limits = c(NA, 800)) +
  labs(
    title = "Economists with Opinions About Differential Privacy Appear Skeptical of\nDifferential Privacy",
    subtitle = "Text: Differential privacy/formal privacy is an approach to protecting privacy that distorts the results more than\nnecessary given the actual risks",
    x = NULL,
    y = NULL,
    caption = "212 respondents did not answer"
  ) +
  theme(panel.grid = element_blank())

```

### Question 8

Q3_8 - Differential privacy/formal privacy is a needed approach to preserve privacy in the face of expanding disclosure risks that current methods are unable to address.

```{r q3.8_unweighted}
aea %>%
  count(Q3_8) %>%
  filter(n != 215) %>%
  mutate(Q3_8 = fct_rev(Q3_8)) %>%
  ggplot(aes(n, Q3_8)) +
  geom_col() +
  geom_text(aes(label = n, hjust = -0.5)) +  
  scale_x_continuous(limits = c(NA, 800)) +
  labs(
    title = "Economists with Opinions About Differential Privacy Appear Skeptical of\nDifferential Privacy",
    subtitle = "Text: Differential privacy/formal privacy is a needed approach to preserve privacy in the face of expanding\ndisclosure risks that current methods are unable to address",
    x = NULL,
    y = NULL,
    caption = "215 respondents did not answer"
  ) +
  theme(panel.grid = element_blank())

```

## Section 4

### Question 1

* What is the largest percentage of significance mismatch you would accept before...?

```{r}
count(aea, is.na(Q4_1_1))

count(aea, is.na(Q4_1_2))

probs <- c(0, 0.05, 0.25, 0.5, 0.75, 0.95, 1)

aea %>%
  summarize(
    probs = probs,
    sig_mismatch = quantile(Q4_1_1, probs = probs, na.rm = TRUE)
  )

aea %>%
  summarize(
    probs = probs,
    sig_mismatch = quantile(Q4_1_2, probs = probs, na.rm = TRUE)
  )

```

```{r q4.1_unweighted}
ggplot() +
  geom_point(
    data = aea,
    mapping = aes(Q4_1_1 / 100, Q4_1_2 / 100), 
    alpha = 0.05
  ) +
  geom_point(
    data = summarize(aea, Q4_1_1 = median(Q4_1_1, na.rm = TRUE), Q4_1_2 = median(Q4_1_2, na.rm = TRUE)),
    mapping = aes(Q4_1_1 / 100, Q4_1_2 / 100), 
    alpha = 0.5,
    color = "#ec008b"
  ) +  
  scale_x_continuous(breaks = seq(0, 1, 0.1), labels = scales::percent_format(accuracy = 1)) +
  scale_y_continuous(breaks = seq(0, 1, 0.1), labels = scales::percent_format(accuracy = 1)) +
  scatter_grid() +
  coord_equal() +
  labs(
    title = "What is the largest percentage of significance mismatch you would\naccept before...",
    subtitle = "Median responses in magenta",
    x = "sacrificing access to the administrative data",
    y = "adversely responding to a journal submission",
    caption = "561 respondents did not answer"
  )

```

### Question 2

What is the largest percentage of sign mismatch you would accept before...?

*This is more conservative than significance, which seems good.*

```{r}
count(aea, is.na(Q4_2_1))

count(aea, is.na(Q4_2_2))

probs <- c(0, 0.05, 0.25, 0.5, 0.75, 0.95, 1)

aea %>%
  summarize(
    probs = probs,
    sign_mismatch = quantile(Q4_2_1, probs = probs, na.rm = TRUE)
  )

aea %>%
  summarize(
    probs = probs,
    sign_mismatch = quantile(Q4_2_2, probs = probs, na.rm = TRUE)
  )

```

```{r q4.2_unweighted}
ggplot() +
  geom_point(
    data = aea,
    mapping = aes(Q4_2_1 / 100, Q4_2_2 / 100), 
    alpha = 0.05
  ) +
  geom_point(
    data = summarize(aea, Q4_2_1 = median(Q4_2_1, na.rm = TRUE), Q4_2_2 = median(Q4_2_2, na.rm = TRUE)),
    mapping = aes(Q4_2_1 / 100, Q4_2_2 / 100), 
    alpha = 0.5,
    color = "#ec008b"
  ) +  
  scale_x_continuous(breaks = seq(0, 1, 0.1), labels = scales::percent_format(accuracy = 1)) +
  scale_y_continuous(breaks = seq(0, 1, 0.1), labels = scales::percent_format(accuracy = 1)) +  
  scatter_grid() +
  coord_equal() +
  labs(
    title = "What is the largest percentage of sign mismatch you would accept\nbefore...",
    subtitle = "Median responses in magenta",    
    x = "sacrificing access to the administrative data",
    y = "adversely responding to a journal submission",
    caption = "563 respondents did not answer"
  )

```

### Question 3

What is the highest amount of absolute relative bias you would accept in an estimate before...?

```{r}
count(aea, is.na(Q4_3_1))

count(aea, is.na(Q4_3_2))

probs <- c(0, 0.05, 0.25, 0.5, 0.75, 0.95, 1)

aea %>%
  summarize(
    probs = probs,
    abs_relative_error = quantile(Q4_3_1, probs = probs, na.rm = TRUE)
  )

aea %>%
  summarize(
    probs = probs,
    abs_relative_error = quantile(Q4_3_2, probs = probs, na.rm = TRUE)
  )

```

```{r q4.3_unweighted}
ggplot() +
  geom_point(
    data = aea,
    mapping = aes(Q4_3_1 / 100, Q4_3_2 / 100), 
    alpha = 0.05
  ) +
  geom_point(
    data = summarize(aea, Q4_3_1 = median(Q4_3_1, na.rm = TRUE), Q4_3_2 = median(Q4_3_2, na.rm = TRUE)),
    mapping = aes(Q4_3_1 / 100, Q4_3_2 / 100), 
    alpha = 0.5,
    color = "#ec008b"
  ) +  
  scale_x_continuous(
    breaks = seq(0, 1.6, 0.2), 
    expand = expansion(mult = c(0, 0.15)), 
    labels = scales::percent_format(accuracy = 1)
  ) +
  scale_y_continuous(
    breaks = seq(0, 1.6, 0.2), 
    expand = expansion(mult = c(0, 0.15)), 
    labels = scales::percent_format(accuracy = 1)
  ) +  
  scatter_grid() +
  coord_equal() +
  labs(
    title = "What is the highest amount of absolute relative bias you would\naccept in an estimate before...",
    subtitle = "Median responses in magenta",    
    x = "sacrificing access to the administrative data",
    y = "adversely responding to a journal submission",
    caption = "578 respondents did not answer"
  )

```

### Question 4

What is the largest confidence interval ratio you would accept in an estimate before...?

```{r}
count(aea, is.na(Q4_4_1))

count(aea, is.na(Q4_4_2))

probs <- c(0, 0.05, 0.25, 0.5, 0.75, 0.95, 1)

aea %>%
  summarize(
    probs = probs,
    ci_ratio = quantile(Q4_4_1, probs = probs, na.rm = TRUE)
  )

aea %>%
  summarize(
    probs = probs,
    ci_ratio = quantile(Q4_4_2, probs = probs, na.rm = TRUE)
  )

```

```{r q4.4_unweighted}
ggplot() +
  geom_point(
    data = aea,
    mapping = aes(Q4_4_1, Q4_4_2), 
    alpha = 0.05
  ) +
  geom_point(
    data = summarize(aea, Q4_4_1 = median(Q4_4_1, na.rm = TRUE), Q4_4_2 = median(Q4_4_2, na.rm = TRUE)),
    mapping = aes(Q4_4_1, Q4_4_2), 
    alpha = 0.5,
    color = "#ec008b"
  ) +  
  scale_x_continuous(limits = c(0, 10)) +
  scale_y_continuous(limits = c(0, 10)) +
  scatter_grid() +
  coord_equal() +
  labs(
    title = "What is the largest confidence interval ratio you would accept in\nan estimate before",
    subtitle = "Median responses in magenta",    
    x = "sacrificing access to the administrative data",
    y = "adversely responding to a journal submission",
    caption = "586 respondents did not answer"
  )

```

## Section 5

### Question 1

5.1: Suppose you gain access to administrative data for regression analysis, but your access is constrained by a privacy budget. How would you spend your privacy budget from the following choices?

```{r q5.1_unweighted}
aea %>%
  count(Q5_1) %>%
  filter(n != 502) %>%
  mutate(Q5_1 = fct_recode(
    Q5_1,
    `One regression specification with moderate noise\nand five robustness checks with more noise` =
      "One regression specification with moderate noise and five robustness checks with more   noise")
  ) %>%  
  mutate(Q5_1 = fct_rev(Q5_1)) %>%
  ggplot(aes(n, Q5_1)) +
  geom_col() +
  geom_text(aes(label = n), hjust = -0.5) +
  scale_x_continuous(limits = c(NA, 400)) +
  labs(
    title = "Economists Want the Goldilocks Budget",
    subtitle = "Text: Suppose you gain access to administrative data for regression analysis, but your access is constrained by a\nprivacy budget. How would you spend your privacy budget from the following choices?",
    x = NULL,
    y = NULL,
    caption = "502 respondents did not answer"
  ) +
  theme(panel.grid = element_blank())

```

## Software

```{r}
sessionInfo()

```

## Conclusions

* The respondents look like members of the AEA. 
* Linear regression and merging data are essential methods. 
* Estimated coefficients and standard errors are essential information from regression models. 
* Most economists are unfamiliar with differential privacy and formal privacy.
* The economists who are familiar with differential privacy are skeptical of differential privacy.
* Economists want very modest amounts of error from formally private methods. 
* Economists seem to have similar standards for the error metrics regardless of if it is their own work or someone else's refereed work. 
* Economists want the Goldilocks budget
